Extract gate/media routes out of agent-channel.ts (Wave D batch 6)
Moves GET /gate and GET /media into agent-channel-gate-media.ts, the
last extractable route cluster in agent-channel.ts (2643 -> 2291
lines). Everything left in the file is composition-root state or
routes tightly entangled with it.
